# Declare frozen_string_literal pragma