--- a/lib/hamlit/attribute_builder.rb
+++ b/lib/hamlit/attribute_builder.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 require 'hamlit/hamlit'
 require 'hamlit/object_ref'
 require 'hamlit/utils'