--- a/lib/hamlit/cli.rb
+++ b/lib/hamlit/cli.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 require 'hamlit'
 require 'thor'