--- a/lib/hamlit/compiler/doctype_compiler.rb
+++ b/lib/hamlit/compiler/doctype_compiler.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 module Hamlit
   class Compiler
     class DoctypeCompiler