--- a/lib/hamlit/compiler/silent_script_compiler.rb
+++ b/lib/hamlit/compiler/silent_script_compiler.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 module Hamlit
   class Compiler
     class SilentScriptCompiler