--- a/lib/hamlit/engine.rb
+++ b/lib/hamlit/engine.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 require 'temple'
 require 'hamlit/parser'
 require 'hamlit/compiler'