--- a/lib/hamlit/error.rb
+++ b/lib/hamlit/error.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 module Hamlit
   class Error < StandardError
     attr_reader :line