--- a/lib/hamlit/escapable.rb
+++ b/lib/hamlit/escapable.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 require 'hamlit/utils'
 
 module Hamlit