--- a/lib/hamlit/filters.rb
+++ b/lib/hamlit/filters.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 require 'hamlit/filters/base'
 require 'hamlit/filters/text_base'
 require 'hamlit/filters/tilt_base'