--- a/lib/hamlit/filters/coffee.rb
+++ b/lib/hamlit/filters/coffee.rb
@@ -1,12 +1,13 @@
+# frozen_string_literal: true
 module Hamlit
   class Filters
     class Coffee < TiltBase
       def compile(node)
         require 'tilt/coffee' if explicit_require?
         temple = [:multi]
-        temple << [:static, "<script>\n".freeze]
+        temple << [:static, "<script>\n"]
         temple << compile_with_tilt(node, 'coffee', indent_width: 2)
-        temple << [:static, "</script>".freeze]
+        temple << [:static, "</script>"]
         temple
       end
     end