--- a/lib/hamlit/filters/erb.rb
+++ b/lib/hamlit/filters/erb.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 module Hamlit
   class Filters
     class Erb < TiltBase