--- a/lib/hamlit/filters/escaped.rb
+++ b/lib/hamlit/filters/escaped.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 module Hamlit
   class Filters
     class Escaped < Base