--- a/lib/hamlit/filters/plain.rb
+++ b/lib/hamlit/filters/plain.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 require 'hamlit/string_splitter'
 
 module Hamlit