--- a/lib/hamlit/filters/preserve.rb
+++ b/lib/hamlit/filters/preserve.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 module Hamlit
   class Filters
     class Preserve < Base