--- a/lib/hamlit/filters/ruby.rb
+++ b/lib/hamlit/filters/ruby.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 module Hamlit
   class Filters
     class Ruby < Base