--- a/lib/hamlit/filters/sass.rb
+++ b/lib/hamlit/filters/sass.rb
@@ -1,12 +1,13 @@
+# frozen_string_literal: true
 module Hamlit
   class Filters
     class Sass < TiltBase
       def compile(node)
         require 'tilt/sass' if explicit_require?
         temple = [:multi]
-        temple << [:static, "<style>\n".freeze]
+        temple << [:static, "<style>\n"]
         temple << compile_with_tilt(node, 'sass', indent_width: 2)
-        temple << [:static, "</style>".freeze]
+        temple << [:static, "</style>"]
         temple
       end
     end