--- a/lib/hamlit/filters/scss.rb
+++ b/lib/hamlit/filters/scss.rb
@@ -1,12 +1,13 @@
+# frozen_string_literal: true
 module Hamlit
   class Filters
     class Scss < TiltBase
       def compile(node)
         require 'tilt/sass' if explicit_require?
         temple = [:multi]
-        temple << [:static, "<style>\n".freeze]
+        temple << [:static, "<style>\n"]
         temple << compile_with_tilt(node, 'scss', indent_width: 2)
-        temple << [:static, "</style>".freeze]
+        temple << [:static, "</style>"]
         temple
       end
     end