--- a/lib/hamlit/filters/text_base.rb
+++ b/lib/hamlit/filters/text_base.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 module Hamlit
   class Filters
     class TextBase < Base