--- a/lib/hamlit/force_escapable.rb
+++ b/lib/hamlit/force_escapable.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 require 'hamlit/escapable'
 
 module Hamlit