--- a/lib/hamlit/helpers.rb
+++ b/lib/hamlit/helpers.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 module Hamlit
   module Helpers
     extend self