--- a/lib/hamlit/identity.rb
+++ b/lib/hamlit/identity.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 module Hamlit
   class Identity
     def initialize