--- a/lib/hamlit/object_ref.rb
+++ b/lib/hamlit/object_ref.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 module Hamlit
   module ObjectRef
     class << self