--- a/lib/hamlit/rails_helpers.rb
+++ b/lib/hamlit/rails_helpers.rb
@@ -1,4 +1,4 @@
-# frozen_string_literal: true
+# frozen_string_literal: false
 require 'hamlit/helpers'
 
 # Currently this Hamlit::Helpers depends on