--- a/lib/hamlit/rails_template.rb
+++ b/lib/hamlit/rails_template.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 require 'temple'
 require 'hamlit/engine'
 require 'hamlit/rails_helpers'