--- a/lib/hamlit/railtie.rb
+++ b/lib/hamlit/railtie.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 require 'rails'
 
 module Hamlit