--- a/lib/hamlit/ruby_expression.rb
+++ b/lib/hamlit/ruby_expression.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 require 'ripper'
 
 module Hamlit