--- a/lib/hamlit/static_analyzer.rb
+++ b/lib/hamlit/static_analyzer.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 require 'hamlit/ruby_expression'
 
 module Hamlit