--- a/lib/hamlit/template.rb
+++ b/lib/hamlit/template.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 require 'temple'
 require 'hamlit/engine'
 require 'hamlit/helpers'