--- a/lib/hamlit/template.rb
+++ b/lib/hamlit/template.rb
@@ -1,4 +1,4 @@
-# frozen_string_literal: true
+# frozen_string_literal: false
 require 'temple'
 require 'hamlit/engine'
 require 'hamlit/helpers'