--- a/lib/hamlit/utils.rb
+++ b/lib/hamlit/utils.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 require 'hamlit/hamlit'
 
 module Hamlit