--- a/lib/hamlit/version.rb
+++ b/lib/hamlit/version.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 module Hamlit
   VERSION = '2.6.1'
 end