Hamlit is using frozen string literal in its source code to optimize compilation.
Since Ruby 2.3 may be widely used now, I prefer `frozen_string_literal: true` rather than appending `.freeze` to all string literals.