@k0kubun 2016-08-26
No effect for performance. ¯\\\_(ツ)\_/¯

## before
```
=================================================
 Compilation: test/haml/templates/standard.haml
=================================================
Calculating -------------------------------------
  haml v4.1.0.beta.1    48.000  i/100ms
         faml v0.8.1    22.000  i/100ms
       hamlit v2.6.1    18.000  i/100ms
-------------------------------------------------
  haml v4.1.0.beta.1    482.724  i/s (2.072ms) -      2.448k
         faml v0.8.1    220.218  i/s (4.541ms) -      1.122k
       hamlit v2.6.1    179.144  i/s (5.582ms) -    900.000 
Comparison:
  haml v4.1.0.beta.1:      482.7 i/s (2.072ms)
         faml v0.8.1:      220.2 i/s (4.541ms) - 2.19x slower
       hamlit v2.6.1:      179.1 i/s (5.582ms) - 2.69x slower
```

## after
```
=================================================
 Compilation: test/haml/templates/standard.haml
=================================================
Calculating -------------------------------------
  haml v4.1.0.beta.1    47.000  i/100ms
         faml v0.8.1    22.000  i/100ms
       hamlit v2.6.1    17.000  i/100ms
-------------------------------------------------
  haml v4.1.0.beta.1    471.825  i/s (2.119ms) -      2.397k
         faml v0.8.1    223.487  i/s (4.475ms) -      1.122k
       hamlit v2.6.1    177.999  i/s (5.618ms) -    901.000 
Comparison:
  haml v4.1.0.beta.1:      471.8 i/s (2.119ms)
         faml v0.8.1:      223.5 i/s (4.475ms) - 2.11x slower
       hamlit v2.6.1:      178.0 i/s (5.618ms) - 2.65x slower
```