@WaKeMaTTa 2016-08-27
Version of ruby are you using? 

`frozen_string_literal: true` >= Ruby 2.3