@k0kubun 2016-08-27
The result is benchmarked by Ruby 2.3.1 on Travis.
https://travis-ci.org/k0kubun/hamlit/jobs/155382106
https://travis-ci.org/k0kubun/hamlit/jobs/155408830